% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_total.R
\name{direction_total}
\alias{direction_total}
\title{Summarizes upstream and downstream movements on arrays}
\usage{
direction_total(dat, resolution = NULL, start_date = min(dat$date_time),
  end_date = max(dat$date_time))
}
\arguments{
\item{dat}{telemetry dataset created using old_pit, new_pit or array_config}

\item{resolution}{summarize data by year, month, week, day or hour}

\item{start_date}{start date of period of interest, default is first date in dataset}

\item{end_date}{end date of period of interest, default is last date in dataset}
}
\description{
Function summarizes upstream and downstream movements of each fish on each array over a user-defined period of time. Such a function can assist in determining residence time between unique arrays. Data can be summarized by year, month, week, day or hour.
}
\details{
Users can apply the direction_total function to the original dataset created by the old_pit or pit_data function, or use the updated dataset created by the array_config function. Arguments start_date and end_date must be entered as yyyy-mm-dd hh:mm:ss.
}
\examples{
#summarize by month
direction_total(new$all_det, “month”)

#summarize by year with a start date of 2015-11-11 10:30:00
direction_total(new$all_det, “year”, “2015-11-11 10:30:00”)
}

