% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first_last}
\alias{first_last}
\title{Summarizes first and last detections on each antenna}
\usage{
first_last(dat, resolution = NULL, start_date = min(dat$date_time),
  end_date = max(dat$date_time))
}
\arguments{
\item{dat}{telemetry dataset created using old_pit, new_pit or array_config}

\item{resolution}{summarize data by year, month, week, day or hour}

\item{start_date}{start date of period of interest, default is first date in dataset}

\item{end_date}{end date of period of interest, default is last date in dataset}
}
\description{
Function summarizes the first and last detections and the difference in time (in minutes and days) between first and last detections on each antenna over a user-defined period of time. Data can be summarized by year, month, week, day or hour.
}
\details{
Users can apply the first_last function to the original dataset created by the old_pit or new_pit function, or use the updated dataset created by the array_config function. Arguments start_date and end_date must be entered as “yyyy-mm-dd hh:mm:ss”.
}
\examples{
#summarize first and last detections by day with a start date of 2015-10-15 08:00:00
first_last(data, “day”, “2015-10-15 08:00:00”)

#summarize first and last detections by month
first_last(data, “month”)
}

