% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_date_from_string}
\alias{parse_date_from_string}
\title{Parse datafile date.}
\usage{
parse_date_from_string(filename, ...)
}
\arguments{
\item{filename}{Text string giving filename with or without full pathname.}

\item{...}{Extra arguments passed to \code{lubridate::ymd}.}
}
\value{
a date in \code{POSIXct} format.
}
\description{
Parses a data download filename and extracts the date.
}
\details{
Used to extract the data download date from a filename with the following form:  \bold{gipi1_report_2017_06_30_13_52_31.txt}.
}
\section{Dave Fifield}{
NA
}

