% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitdb.R
\name{pitdb_close}
\alias{pitdb_close}
\title{Close a pit tag database}
\usage{
pitdb_close(ch)
}
\arguments{
\item{ch}{A open ODBC channel.}
}
\value{
Nothing
}
\description{
The function will close the connection to the Access PIT tag database via ODBC.
}
\section{Author}{
 Dave Fifield
}

