% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_captures.R
\name{pitdb_extract_captures}
\alias{pitdb_extract_captures}
\title{Extract capture events PIT tag database.}
\usage{
pitdb_extract_captures(pitdb.path = NULL, years = NULL, verbose = FALSE)
}
\arguments{
\item{pitdb.path}{\[character:\sQuote{NULL}]\cr Full path name to the PIT tag
database.}

\item{years}{\[integer:\sQuote{NULL}]\cr Optional. Either a single year or a
vector of two years denoting "from" and "to" (inclusive).}
}
\description{
This function will connect to the Access database, create a
 series of queries and import the desired information in a data frame.
}
\section{Author}{
 Dave Fifield
}

