% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_nest_attempt.R
\name{pitdb_impute_nest_attempt}
\alias{pitdb_impute_nest_attempt}
\title{Parse data from a Raspberry Pi report file}
\usage{
pitdb_impute_nest_attempt(dt, ch)
}
\arguments{
\item{dt}{A dtaframe containing at least fields: \code{Band, Burrow, Date}}

\item{ch}{Open RODBC channel connecting to the database}
}
\value{
Nothing.
}
\description{
The function will create and/or fill in nest attempt records from info on band no, date and burrow.
}
\details{
If no record currently exists in tblNestAttempt for this bird/burrow combination in this year, then a new record is created.
If a record already exists for this bird/year for another burrow then an error message is printed and the function returns.
Otherwise if a record already exists for this bird/year and the \code{Bird1} field is not already assigned to this bird
(which would produce an error) and the \code{Bird2} field is empty then the \code{Bird2} field is filled in with this bird's ID.
}
\section{Author}{
 Dave Fifield
}

