% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_board_deployment.R
\name{pitdb_insert_board_deployment}
\alias{pitdb_insert_board_deployment}
\title{Insert board deployment data}
\usage{
pitdb_insert_board_deployment(dt, ch)
}
\arguments{
\item{dt}{A dataframe containing at least \code{BoardID, Burrow, FromDate}.
Also may contain \code{ToDate, Revision, FlashSize, Status,
Plot, Habitat, BurrowNotes} - see below for details.}

\item{ch}{Open RODBC channel connecting to the database}
}
\value{
Nothing.
}
\description{
The function inserts a single board deployment record into the
database.
}
\details{
Creates appropriate entries in tblBoardDeployment, tblBoard,
 and tblBurrow to record a board being deployed in a burrow from \code{FromDate}
 to \code{ToDate} (which may be NA).

 Date is the date and time of the deployment of the board in the burrow. It must
 be a \code{POSIXct}.

 If \code{BoardID} does not exist in \code{tblBoard} it is added but this requires
 fields for \code{Revision, FlashSize,} and \code{Status}.

 if \code{Burrow} does not exist in \code{tblBurrow} it will be added but
 this requires fields for \code{Plot} and \code{Habitat} and optionally
 \code{BurrowNotes}. See table \code{lkpHabitat} for habitat details.
}
\section{Author}{
 Dave Fifield
}

