% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_burrow_check.R
\name{pitdb_insert_burrow_check}
\alias{pitdb_insert_burrow_check}
\title{Insert burrow check data}
\usage{
pitdb_insert_burrow_check(dt, ch)
}
\arguments{
\item{dt}{A dataframe containing at least \code{Burrow, Status, DateTime}.
May also contain \code{BandNo, CheckNote}. See below for more details.}

\item{ch}{Open RODBC channel connecting to the database}
}
\value{
Nothing.
}
\description{
The function inserts a single burrow check record into the
database.
}
\details{
Creates appropriate entries in tblBurrowCheck, tblBurrow,
 and tblBird to record the status of a burrow on \code{DateTime}.

 DateTime is the date and time of the check. It must
 be a \code{POSIXct}.

 See table \code{lkpBurrowStatus} for valied \code{Status} values.

 If \code{Burrow} does not exist in \code{tblBurrow} it will be added but
 this requires fields for \code{Plot} and \code{Habitat} and optionally
 \code{BurrowNotes}. If the burrow does already exist, these fields are ignored
 if provided. See table \code{lkpHabitat} for habitat details.

 If a bird is found in the burrow, it is indicated by \code{BandNo}.
 If this bird is not already in \code{tblBirds}), then a new record is
 created in \code{tblBirds} but this
 requires fields for \code{AgeAtBanding, Species, and Colony} and optionally
 \code{Sex} and \code{BirdNote}. If the bird is already in \code{tblBirds} then
 these fields are ignored.
}
\section{Author}{
 Dave Fifield
}

