% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_capture.R
\name{pitdb_insert_capture}
\alias{pitdb_insert_capture}
\title{Insert capture data}
\usage{
pitdb_insert_capture(dt, ch)
}
\arguments{
\item{dt}{A dataframe containing at least \code{Band, Burrow, DateTime} and
optionally \code{Colony, Species, Age, TotW, BagW, Tarsus, Wing, BloodCard,
BloodVial, BloodVialAmount, CloSwab, PharSwab, Feathers, SampleNote,
Behaviour, PIT_deployed, PIT_removed}.}

\item{ch}{Open RODBC channel connecting to the database}
}
\value{
Nothing.
}
\description{
The function inserts a single capture record into the database.
}
\details{
Creates or updates appropriate entries in tblBirds, tblBands, tblTags,
 tblNestAttempt, tblCapture, tblMorpho, and tblTagDeployment.

 If this bird is not already in \code{tblBirds}, then a new record is
 created in \code{tblBirds} but this
 requires fields for \code{AgeAtBanding, Species, and Colony} and optionally
 \code{Sex} and \code{BirdNote}. If the bird is already in \code{tblBirds} then
 these fields are ignored.

 If \code{Burrow} does not exist then it is created in \code{tblBurrow} which
 requires field \code{Plot} and optionally \code{Habitat} and \code{BurrowNotes}.

 If \code{Species} or \code{Colony} are missing, they default to "Gull Island"
 and "Leach's Storm-Petrel" respectively. See tables \code{tblSpeciesInfo} and
 \code{lkpColony} respectively for acceptable values.

 If \code{PIT_deployed} tag does not already
 exist in the database, then \code{dt} must contain fields \code{Manufac}
 and \code{TagType} - see tables \code{lkpManufacturers} and \code{lkpTagTypes}
 in the database for acceptable values.

 If \code{PIT_removed} is supplied then it should exist in
 \code{tblTagDeployment}, otherwise a warning is displayed and no removal
 is recorded.

 \code{Behaviour} refers to behaviour when arm is inserted into burrow and includes
 things like "run to back", "bite once", "bite multiple times", "vocalize",
 "calm/docile/neutral"
}
\section{Author}{
 Dave Fifield
}

