% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chick_banding.R
\name{pitdb_insert_chick_banding}
\alias{pitdb_insert_chick_banding}
\title{Import chick banding}
\usage{
pitdb_insert_chick_banding(dt, ch)
}
\arguments{
\item{dt}{A dataframe containing at least fields: \code{Band, Burrow, Date, TotW, BagW, Tarsus, PIT, Contents}}

\item{ch}{Open RODBC channel connecting to the database}
}
\value{
Nothing.
}
\description{
The function import a single chick banding record to the database.
}
\details{
Creates appropriate entries in tblBirds, tblNestAttempt, tblCapture, tblMorpho, tblTagDeployment, tblBurrowCheck
}
\note{
Currently assumes that all birds are at Gull Island, tags starting with 062 are Cyntag 10mm tags, otherwise they are CoreRFID 12mm tags.
}
\section{Author}{
 Dave Fifield
}

