% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{pitdb_load_files}
\alias{pitdb_load_files}
\title{Load multiple PIT tag datafiles into a database.}
\usage{
pitdb_load_files(ch, filenames, ...)
}
\arguments{
\item{ch}{(required) Channel to open database returned from \link{pitdb_open}.}

\item{filenames}{(required) Character vector of full pathnames to files
containing data from a monitor board.}

\item{...}{Other args passed to \link{pitdb_load_file}}
}
\value{
Nothing
}
\description{
Repeatedly calls \link{pitdb_load_file} to load multiple files
 into a Microsoft Access database.
}
\details{
This calls \link{pitdb_load_file} once for each element of \code{filenames}
 in order to read a text file of records that have been extracted
 from a monitor board and import them into the database indicated by
 \code{channel}. These files are typically downloaded via direct cable
 connection or received from a Raspberry Pi base station.
}
\section{Author}{
 Dave Fifield
}

