% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bird_report_file.R
\name{pitdb_parse_bird_report_file}
\alias{pitdb_parse_bird_report_file}
\title{Parse data from a Raspberry Pi report file}
\usage{
pitdb_parse_bird_report_file(filename, verbose = FALSE)
}
\arguments{
\item{filename}{Path to file containing data dumped from a monitor board.}

\item{verbose}{Print information regarding numbers of status records, tag read
records, mark_upload records, # of boards reporting, and ghost reads.}
}
\value{
Returns a list with elements \code{tag_reads, statuses, uploads, bad_recs}. Individual elements of the list will be
null if \code{filename} doesn't contain any records of that type.
}
\description{
The function will parse a bird report datafile.
}
\details{
This function reads a text file of records that have been extracted
 from a monitor board. These files are typically downloaded via direct cable
 connection or via email from a Raspberry Pi base station.

 In most cases \code{filename} will contain info from multiple monitor
 boards, e.g., the daily files emailed from a Raspberry Pi base station.

 Does not attempt to further split tag_reads into normal reads, webserver tags, ghost tags, etc.
 since this requires info on known tags/prefixes, etc from the database and it is desirable to
 have this function work whether database is available or not.

 talk about format of filename somewhere

 deal with different data formats in 2016 and 2017

 deal with conversion of voltages and MCU temperature

 prints a summary of what was done (consider how this is done: jsut via print
 or does it return a summary object which has a print method...probably just
 print it - easier).

 deal with "Bad record" rows
}
\section{Author}{
 Dave Fifield
}

