% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bird_report_file.R
\name{pitdb_process_bird_report_file}
\alias{pitdb_process_bird_report_file}
\title{Parse data from a Raspberry Pi report file}
\usage{
pitdb_process_bird_report_file(filename, sp_tags = NULL, verbose = F)
}
\arguments{
\item{filename}{Path to file containing data dumped from a monitor board.}

\item{sp_tags}{A dataframe that indicates the identity of test tags, known
 (i.e., deployed) tag prefixes, and webserver_mode tag prefixes.

 The dataframe must contain at least two columns: "Typ" and "Val", Where Typ has entries for at least
 \code{test_tag, web_prefix, known_prefix}.}

\item{verbose}{Print information regarding numbers of status records, tag read
records, mark_upload records, # of boards reporting, and ghost reads.}
}
\value{
Returns TRUE on success and FALSE on error.
}
\description{
The function will parse a bird report datafile.
}
\details{
This function reads a text file of records that have been extracted
 from a monitor board. These files are typically downloaded via direct cable
 connection or via email from a Raspberry Pi base station.

 In most cases \code{inputfile} will contain info from multiple monitor
 boards, e.g., the daily files emailed from a Raspberry Pi base station.

 talk about format of inputfile somewhere

 deal with different data formats in 2016 and 2017

 deal with conversion of voltages and MCU temperature

 prints a summary of what was done (consider how this is done: jsut via print
 or does it return a summary object which has a print method...probably just
 print it - easier).

 deal with "Bad record" rows
}
\section{Author}{
 Dave Fifield
}

