% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_parsed_file.R
\name{pitdb_summarize_parsed_file}
\alias{pitdb_summarize_parsed_file}
\title{Print a summary of data parsed from a Raspberry Pi report file}
\usage{
pitdb_summarize_parsed_file(dat, ch = NULL, verbose = FALSE)
}
\arguments{
\item{dat}{List (typically returned from \code{pitdb_parse_bird_report_file})
containing elements \code{tag_reads, statuses, uploads, bad_recs}.}

\item{ch}{an open connection to the database. If not null, tag_read records are further categorized
as known_tags, webserver tags, test tags, and unknown tags.}

\item{verbose}{If \code{TRUE}, print details of each record.}
}
\description{
The function summarizes data parsed from a bird report datafile.
}
\details{
This function takes the output from \code{pitdb_parse_bird_report_file}
 and summarizes the numbers of each type of info contained in the parsed file.
}
\section{Author}{
 Dave Fifield
}

