% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bird_report_file.R
\name{pitdb_summary_parsed_file}
\alias{pitdb_summary_parsed_file}
\title{Print a summary of data parsed from a Raspberry Pi report file}
\usage{
pitdb_summary_parsed_file(dat, show_data = FALSE)
}
\arguments{
\item{dat}{List (typically returned from \code{pitdb_parse_bird_report_file}
containing elements \code{status, uploads, tag_reads, bad_recs}.
If optional elements \code{one_read, fakeGhost_reads, test_reads, web_reads} are included
they will be reported on too.}

\item{sp_tags}{A dataframe that indicates the identity of test tags, known
 (i.e., deployed) tag prefixes, and webserver_mode tag prefixes.

 The dataframe must contain at least two columns: "Typ" and "Val", Where Typ has entries for at least
 \code{test_tag, web_prefix, known_prefix}.}
}
\description{
The function will summarize data parsed from a a bird report datafile.
}
\details{
This function takes the output from \code{pitdb_parse_bird_report_file} optionally
 augmented by elements further characterizing  \code{tag_reads} (i.e.,  \code{one_read, fakeGhost_reads, test_reads, web_reads})
 and summarizes the numbers of each type of info contained in the parsed file.
}
\section{Author}{
 Dave Fifield

############# UNFINISHED ####################
}

