% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitr.R
\docType{package}
\name{pitr}
\alias{pitr}
\title{pitr: A package for manipulating data from PIT tag monitors.}
\description{
blah blah blah
}
\details{
Data formats for PIT tag monitor boards.

 In most cases \code{inputfile} will contain info from multiple monitor
 boards, e.g., the daily files emailed from a Raspberry Pi base station.

 talk about format of inputfile somewhere

 deal with different data formats in 2016 and 2017
}
\section{Author}{
 Dave Fifield
}

