% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_handler.R
\name{pitr_setup_git_creds}
\alias{pitr_setup_git_creds}
\title{Setup credentials for connecting with GitHub PIT tag data repos}
\usage{
pitr_setup_git_creds(username = "gull-island")
}
\arguments{
\item{username}{Character string containing the github username. (Default:
"gull-island").}
}
\value{
credentials created by \code{git2r::cred_user_pass}
}
\description{
This function looks up the GITHUB_PAT environment variable and
 combines it with \code{username} to create and return credentials suitable
 for passing to \code{git2r::pull()}, etc.
}
\details{
Since this function uses GitHub Personal Access Tokens, the returned
 credentials will only work with repos cloned with HTTPS.
}
\section{Author}{
 Dave Fifield
}

