% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_handler.R
\name{pitr_setup_plot_repos}
\alias{pitr_setup_plot_repos}
\title{Clone data repos for a given year to the local machine from GitHub}
\usage{
pitr_setup_plot_repos(
  base_folder,
  year,
  plot_nos = 1:6,
  force_delete = FALSE,
  user.name = "gull-island",
  user.email = "gull-island@gmail.com"
)
}
\arguments{
\item{base_folder}{The folder under which individual plot data repos are cloned.}

\item{year}{The year to setup repos for.}

\item{plot_nos}{A vector of plot numbers indicating which plot repos to set up.}

\item{force_delete}{Should an existing local folder for this repo be deleted
if found? Defalut:\code{FALSE}}
}
\value{
xxx
}
\description{
Sets up local data repos for all plots in a given year and
   clones them from GitHub
}
\details{
If an existing local folder is found for any repo and force_delete is
   \code{FALSE}, then a warning is generated for that repo and now further action
   will be taken for it.
}
\section{Author}{
 Dave Fifield
}

