% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_handler.R
\name{pull_from_repo}
\alias{pull_from_repo}
\title{Pull latest PIT tag data from Git repositories}
\usage{
pull_from_repo(repos, cred = NULL)
}
\arguments{
\item{repos}{A character vector of one or more complete pathnames to a local
repository. Repositories must have a remote defined in order for the pull
to succes}

\item{cred}{Credentials as returned by pitr_setup_git_crds() (default: \code{NULL}).}
}
\value{
Nothing.
}
\description{
This function executes a 'git pull' (using \code{git2r::pull()})
 for each repository listed in repos.
}
\details{
Attempts to pull any updates from the remote for each repository given
 in \code{repos}. Prints a information message for each repo giving the location of
 both the local and remote.
}
\section{Author}{
 Dave Fifield
}

