% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_handler.R
\name{setup_git_creds}
\alias{setup_git_creds}
\title{Setup credentials for connecting with GitHub PIT tag data repos}
\usage{
setup_git_creds(username = "gull-island", pat = NULL)
}
\arguments{
\item{username}{Character string containing the github username. (Default:
"gull-island").}

\item{pat}{Character string containing the Personal Access Token (PAT). If
\code{NULL} (default) it is looked up in the \code{GITHUB_PAT} environment
variable.}
}
\value{
Credentials created by \code{git2r::cred_user_pass}
}
\description{
This function creates and returns credentials suitable
 for passing to \code{git2r::pull()}, etc.
}
\details{
Since this function uses GitHub Personal Access Tokens, the returned
 credentials will only work with repos cloned with HTTPS. Note that if this
 function is called from a process that is started via Windows Task Scheduler
 under the "SYSTEM" pseudo-user, then the \code{GITHUB_PAT} environment
 variable will not be set. In such a case, you should pass the PAT via the
 \code{pat} argument.
}
\section{Author}{
 Dave Fifield
}

