% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volt_plot.R
\name{volt_plot}
\alias{volt_plot}
\title{Plots voltage of each reader over the study duration}
\usage{
volt_plot(volt_dat, file_path = getwd())
}
\arguments{
\item{volt_dat}{voltage dataset created using \code{\link{new_pit}} function}

\item{file_path}{filepath to save the voltage plot}
}
\description{
Function plots the voltage of each reader over the study duration.
}
\details{
Plot is saved in the user-specified working directory.
}
\examples{
#save voltage plot to a Dropbox folder
volt_plot(dat$volt_dat, "~/Dropbox (InStream)/Projects/62 - PIT R and D/4 – Figures and Tables")
}

