% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_config.R
\name{array_config}
\alias{array_config}
\title{Restructure the configuration of arrays}
\usage{
array_config(dat, configuration, arrayname = NULL, r1 = NULL, r2 = NULL,
  r3 = NULL, r4 = NULL, readername = NULL, newreader1antennas = NULL,
  ao1 = NULL, ao2 = NULL, ao3 = NULL, ao4 = NULL, an1 = NULL,
  an2 = NULL, an3 = NULL, an4 = NULL)
}
\arguments{
\item{dat}{telemetry dataset created using old_pit or new_pit function}

\item{configuration}{either ‘combine’, ‘split’ or ‘rename_antennas’}

\item{arrayname}{unique name of an array}

\item{r1}{name of reader 1}

\item{r2}{name of reader 2}

\item{r3}{name of reader 3}

\item{r4}{name of reader 4}

\item{readername}{name of reader to split}

\item{newreader1antennas}{specific antennas to be grouped into reader 1}

\item{ao1}{old antenna 1}

\item{ao2}{old antenna 2}

\item{ao3}{old antenna 3}

\item{ao4}{old antenna 4}

\item{an1}{new antenna 1}

\item{an2}{new antenna 2}

\item{an3}{new antenna 3}

\item{an4}{new antenna 4}
}
\value{
An updated dataset for further analysis using det_eff, direction, direction_total and first_last functions.
}
\description{
Function allows users to combine unique readers into an array, split readers with multiple antennas into single antennas, and rename up to four antennas on one reader or one array. Use of this function allows users to manage data for further analysis using det_eff, direction, direction_total, and first_last functions.
}
\details{
Function is dependent on what the user defines in the ‘configuration’ argument. Argument ‘combine’ is used to combine up to four single readers into one array. If a user wants to combine a multi reader with another reader (either a multi reader or single reader), the user first must split the multi reader into two or more single readers and then combine them together into an array. Argument ‘split’ is used to split multi readers into two or more single readers to allow the combination of single readers into a user-specified array. Argument ‘new_reader_1_antennas’ is used to define the antennas to group into reader 1. All other antennas are grouped into reader 2. Argument ‘rename_antennas’ is used to either rename antennas part of an array (if array_name is specified) or antennas part of a reader (if reader_name is specified). Users can rename up to four antennas for one reader or one array. Arguments with the prefix ‘ao’ correspond to the old (or current) antenna numbering scheme, whereas arguments with the prefix ‘an’ correspond to the new antenna numbering scheme. Old antennas must be specified in order within the array_config function. Depending on the setup of the study, the user may have to run the array_config function several times. Note that use of arguments ‘combine’, ‘split’ and ‘rename_antennas’ must be iterative (see Examples). Ultimately the array_config function outputs an updated dataset to be used for further analysis.
}
\examples{
#combine two single readers into an array called ‘bridge_reach34’
array_config(new$all_det, configuration = "combine", array_name = "bridge_reach34", r1 = "bridge_reach34_ant1", r2 = "bridge_reach34_ant2")

#split the ‘bridge_counter’ multi reader into two single readers; include A2 into reader #1
array_config(new$all_det, configuration = "split", reader_name = "bridge_counter", new_reader_1_antennas = c(2))
#or
array_config(new$all_det, configuration = "split", reader_name = "bridge_counter", new_reader_1_antennas = 2)

#within the ‘bridge_counter’ array, rename the old A1 to A2 and the old A2 to A1
array_config(new$all_det, configuration = "rename_antennas", array_name = "bridge_counter", ao1 = 2, ao2 = 1, an1 = 1, an2 = 2)
#within the ‘bridge_reach34_ant2’ reader, rename the old A2 to A1
array_config(new$all_det, configuration = "rename_antennas", reader_name = "bridge_reach34_ant2", ao1 = 2, an1 = 1)
}

