% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_eff.R
\name{det_eff}
\alias{det_eff}
\title{Computes the detection efficiency of individual antennas}
\usage{
det_eff(dat, resolution = NULL, direction, start_date = min(dat$date_time),
  end_date = max(dat$date_time))
}
\arguments{
\item{dat}{telemetry dataset created using old_pit, new_pit or array_config}

\item{resolution}{summarize data by year, month, week, day or hour}

\item{direction}{user-specified direction of fish movement}

\item{start_date}{start date of period of interest, default is first date in dataset}

\item{end_date}{end date of period of interest, default is last date in dataset}
}
\description{
Function that computes the detection efficiency of individual antennas based on the array configuration and the user’s assumption of the direction of fish movement (up, down or resident). Data can be summarized by year, month, week, day or hour.
}
\details{
Users can apply the det_eff function to the original dataset created by the pit_data function, or use the updated dataset created by the array_config function. Arguments start_date and end_date, if specified, must be entered as yyyy-mm-dd hh:mm:ss.
}
\examples{
#compute by month for fish assumed to be moving upstream
det_eff(new$all_det, "month", "up")

#compute by week for fish assumed to be moving upstream with a start date of 2015-10-11 08:45:00
det_eff(new$all_det, "week", "up", start_date = "2015-10-11 08:45:00")

#compute by day for fish assumed to be moving downstream
det_eff(new$all_det, "day", "down")

#compare by month for fish assumed to be resident
det_eff(new$all_det, "month", "resident")
}

