% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Computes the direction of movement}
\usage{
direction(dat)
}
\arguments{
\item{dat}{telemetry dataset created using \code{\link{old_pit}}, \code{\link{new_pit}} or \code{\link{array_config}}}
}
\value{
A dataframe summarizing the direction of movement.
}
\description{
Function that determines the direction of movement if there were two or more antennas deployed in the study.
}
\details{
User can apply the direction function to the original dataset created by the \code{\link{old_pit}} or \code{\link{new_pit}} function, or use the updated dataset created by the \code{\link{array_config}} function.
}
\examples{
#determine the direction of fish movement
direction(new$all_det)
}

