% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first_last}
\alias{first_last}
\title{Summarizes first and last detections on individual antennas}
\usage{
first_last(data, resolution = NULL, start_date = min(data$date_time),
  end_date = max(data$date_time))
}
\arguments{
\item{data}{telemetry dataset created using \code{\link{old_pit}}, \code{\link{new_pit}} or \code{\link{array_config}}}

\item{resolution}{summarize data by year, month, week, day or hour}

\item{start_date}{start date of period of interest, default is first date in dataset}

\item{end_date}{end date of period of interest, default is last date in dataset}
}
\value{
A dataframe summarizing the first and last detections on individual antennas.
}
\description{
Function summarizes the first and last detections and the difference in time (in minutes and days) between first and last detections on each antenna over a user-defined period of time. Data can be summarized by year, month, week, day or hour.
}
\details{
Users can apply the \code{first_last} function to the original dataset created by the \code{\link{old_pit}} or \code{\link{new_pit}} function, or use an updated dataset created by the \code{\link{array_config}} function. Arguments \code{start_date} and \code{end_date}, if specified, must be entered as “yyyy-mm-dd hh:mm:ss”.
}
\examples{
#load test dataset
oregon_rfid <- new_pit(data = "oregon_rfid", test_tags = NULL, print_to_file = FALSE, time_zone = "America/Vancouver")

#summarize first and last detections by day with a start date of 2015-10-15 08:00:00
first_last(oregon_rfid, “day”, “2015-10-15 08:00:00”)

#summarize first and last detections by month
first_last(oregon_rfid, “month”)
}
