% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bird_report_file.R
\name{pitdb_load_file}
\alias{pitdb_load_file}
\title{Load a PIT tag datafile into a database.}
\usage{
pitdb_load_file(ch = NULL, filename = NULL, date_ = NULL,
  fetch_type = "WiFi", ignore_test_board = TRUE, test_board_ID = 1,
  detect_non_reporters = TRUE, parse_summary = FALSE, verbose = FALSE)
}
\arguments{
\item{ch}{Channel to open database returned from \code{pitdb.open}.}

\item{filename}{Path to file containing data dumped from a monitor board.}

\item{date_}{Date on which the data was downloaded. Normally this is parsed
from the \code{filename} but it can be over-ridden with this argument.}

\item{fetch_type}{How the data was fetched from the board. Valid values are
"WiFi" and "CableConnect".}

\item{ignore_test_board}{Should data from the test board (normally #1) be ignored. Default = \code{TRUE}.}

\item{test_board_ID}{The board ID of the test board (normally 1). Used to ignore testing data.}

\item{detect_non_reporters}{Produce warning message for any deployed boards
not reporting data in \code{filename}.
\code{Sanity checks} below.}

\item{parse_summary}{Print a summary of the parsed data file (default = \code{FALSE})}

\item{verbose}{(default \code{FALSE}) show inserted data when insert_summary is \code{TRUE}}
}
\value{
Returns TRUE on success and FALSE on error.
}
\description{
The function will load a datafile downloaded from a PIT tag
 monitor board into a Microsoft Access database.
}
\details{
This function reads a text file of records that have been extracted
 from a monitor board and imports them into the database indicated by
 \code{channel}. These files are typically downloaded via direct cable
 connection or via email from a Raspberry Pi base station.

 In most cases \code{filename} will contain info from multiple monitor
 boards, e.g., the daily files emailed from a Raspberry Pi base station. In
 this case, it is helpful to know which deployed boards failed to send their
 data that day. If  \code{detect_non_reporters} = \code{TRUE},
 \code{pitdb_load_file} will issue a warning message listing any boards that
 are currently deployed (as defined in tblBoardDeploy in the Access database)
 but did not provide any data in \code{filename}. The meaning of "currently"
 in the last sentence means "on the date that the data was download" which
 comes from the filename but which can be overridded by \code{date}.

 What it does in the database......

 inserts records one by one in case an insert fails, prints warning regarding
 failed inserts.

 talk about format of filename somewhere

 deal with different data formats in 2016 and 2017

 deal with conversion of voltages and MCU temperature

 prints a summary of what was done (consider how this is done: jsut via print
 or does it return a summary object which has a print method...probably just
 print it - easier).

 Decide what to do about unknown tag IDs (legitimate and ghost reads) since
 the database currently requires all tags in tblTagRead table to exist
 apriori in tblTags. Should they go in some other table, or just be
 discarded. Put them in failed load table.

 talk about creation of a tblImports record.

 deal with "Bad record" rows
}
\section{Sanity checks}{


 \itemize{ \item detects any board not reporting if
 \code{detect_non_reporters} = TRUE. \item issues a warning (and fails to add
 record) if an attempt is made to import a record that already exists. \item
 issues a warning (and fails to add record) if a tag_read record refers to an
 unknown tag ID. This could be either a legitimate tag that has not been
 added to the database or a so-called \emph{ghost read}. Ghost reads are
 almost entirely eliminated as of rfid board software version 1.1, since only
 tags with known prefixes (compiled in) are recorded. However, it is remotely
 possible that a ghost read with a known prefix could occur. The number of
 reads field is always 1 (??) in a ghost read, but rarely so for a true tag
 read. Setting \code{detect_ghost_reads} = \code{TRUE} will issue a warning
 for any reads of valid tag IDs where the number of reads is 1. }
}

\section{Author}{
 Dave Fifield
}

