% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_downloads.R
\name{pitdb_process_data_downloads}
\alias{pitdb_process_data_downloads}
\title{Process downloaded PIT tag data}
\usage{
pitdb_process_data_downloads(db = NULL, path = NULL, start_time = XXXX,
  end_time = XXXX, sleep_time = 10)
}
\arguments{
\item{db}{The pathname to the Microsoft Access database to insert downloaded files into.}

\item{path}{The pathname for the folder where files will magically appear waiting to be processed. Usually placed there my Outlook or some other mailer.}

\item{start_time}{Time of day to start looking for new files. Useful to avoid wasting time looking for new files during parts of the day when they
are not expected to arrive. Default \code{12:00:00 noon}}

\item{end_time}{Time of day to stop looking for new files. Useful to avoid wasting time looking for new files during parts of the day when they
are not expected to arrive. Default \code{14:00:00}}

\item{sleep_time}{Minutes between checks for new files. Default 10 minutes.}
}
\value{
Doesn't normally return.
}
\description{
The function waits for new datafiles to arrive in folder \code{path} and then executes and then calls other functions to insert data
into database
}
\details{

}
\section{Author}{
 Dave Fifield
}

