% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_downloads.R
\name{pitdb_process_data_downloads}
\alias{pitdb_process_data_downloads}
\title{Process downloaded PIT tag data}
\usage{
pitdb_process_data_downloads(
  db = NULL,
  path = NULL,
  report_path = NULL,
  logfile_path = NULL,
  log_level = "INFO",
  compare_full_pathname = FALSE,
  start_time = NULL,
  end_time = NULL,
  sleep_time = NULL
)
}
\arguments{
\item{db}{The pathname to the Microsoft Access database to insert downloaded files into.}

\item{path}{The pathname to the folder where data files wait to be processed. Usually
placed there by git pull. Use C:/foo/bar/blah Unix forward slash notaion.}

\item{report_path}{Full pathname to folder where knitted import reports are to be stored. Defaults to \code{path/../Import records}}

\item{logfile_path}{Full pathname of file to receive logging information. Defaults to \code{dirname(db)/import_log.txt}.}

\item{log_level}{One of \code{logging::loglevels}. E.g. \code{ c("NOTSET", "FINEST", "FINER", "FINE", "DEBUG", "INFO", "WARN", "ERROR", "CRITICAL", "FATAL") }
Default is \code{"INFO"}}

\item{start_time}{Time of day to start looking for new files. Useful to avoid wasting time looking for new files during parts of the day when they
are not expected to arrive. NOT IMPLEMENTED.}

\item{end_time}{Time of day to stop looking for new files. Useful to avoid wasting time looking for new files during parts of the day when they
are not expected to arrive. NOT IMPLEMENTED}

\item{sleep_time}{If \code{NULL} (default), return after importing any new files.
Otherwise, either number of seconds between checks for new files or a
character string of form "H:M:S" defining the time to wake each day.}

\item{compare_full_pathanme}{(Default \code{FALSE}). If \code{TRUE} then comparing filenames
in \code{path} to those already imported into \code{db} (in tblImports)
  considers the full pathname, otherwise just considers the filename.}
}
\value{
Doesn't normally return, but -1 on error.
}
\description{
The function waits for new datafiles to arrive in folder \code{path} and imports them into the database specified by \code{db}.
}
\details{
none as of yet
}
\section{Author}{
 Dave Fifield
}

